Changes in signals: model signal moved to apps.py, added new signal
Because there were serious circular imports and import errors happening,
I decided to move `trainingrequest_m2m_changed` signal from
`workshops.signals` to `workshops.apps`.

A new signal was created for saving comments when someone submits a form
with "comment" field, but the form itself is intended to create another
object. This is the easiest solution I came up with to inject creating
comments when creating other objects easily.
